WIP: Swap get-path back to /api/node
TestDistribFileStore passes; ship it.
