--- solr/api/src/java/org/apache/solr/client/api/model/FileStoreRawFileResponse.java
@@ -1,9 +1,0 @@
-package org.apache.solr.client.api.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class FileStoreRawFileResponse extends SolrJerseyResponse {
-  @JsonProperty("content") // A flag value that RawResponseWriter handles specially
-  public Object
-      output; // Traditionally a 'ContentStream' that the server (via RawResponseWriter) can write
-}
