+++ solr/core/src/java/org/apache/solr/filestore/ClusterFileStore.java
@@ -18,32 +18,23 @@
 package org.apache.solr.filestore;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.apache.solr.handler.ReplicationHandler.FILE_STREAM;
 
 import jakarta.inject.Inject;
 import java.io.IOException;
 import java.io.InputStream;
 import java.lang.invoke.MethodHandles;
 import java.nio.ByteBuffer;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.solr.api.JerseyResource;
 import org.apache.solr.client.api.endpoint.ClusterFileStoreApis;
-import org.apache.solr.client.api.model.FileStoreDirectoryListingResponse;
-import org.apache.solr.client.api.model.FileStoreJsonFileResponse;
-import org.apache.solr.client.api.model.FileStoreMissingFileResponse;
 import org.apache.solr.client.api.model.SolrJerseyResponse;
 import org.apache.solr.client.api.model.UploadToFileStoreResponse;
 import org.apache.solr.common.SolrException;
-import org.apache.solr.common.params.CommonParams;
-import org.apache.solr.common.params.ModifiableSolrParams;
-import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.common.util.StrUtils;
 import org.apache.solr.core.CoreContainer;
-import org.apache.solr.core.SolrCore;
 import org.apache.solr.jersey.PermissionName;
 import org.apache.solr.pkg.PackageAPI;
 import org.apache.solr.request.SolrQueryRequest;
@@ -200,125 +191,6 @@ public class ClusterFileStore extends JerseyResource implements ClusterFileStore
       throw new SolrException(SolrException.ErrorCode.BAD_REQUEST, "jar in use, can't delete");
     }
     doDelete(filePath, localDelete);
-    return response;
-  }
-
-  // TODO - this single "get" operation actually supports several different chunks of functionality:
-  // syncing,
-  //  directory listing, file-fetching, metadata-fetching. We should split it up into multiple
-  // distinct APIs
-  @Override
-  public SolrJerseyResponse getFile(String path, Boolean sync, String getFrom, Boolean meta) {
-    final var response = instantiateJerseyResponse(SolrJerseyResponse.class);
-
-    if (Boolean.TRUE.equals(sync)) {
-      try {
-        fileStore.syncToAllNodes(path);
-        return response;
-      } catch (IOException e) {
-        throw new SolrException(SolrException.ErrorCode.BAD_REQUEST, "Error getting file ", e);
-      }
-    }
-
-    final var pathCopy = path;
-    if (getFrom != null) {
-      coreContainer
-          .getUpdateShardHandler()
-          .getUpdateExecutor()
-          .submit(
-              () -> {
-                log.debug("Downloading file {}", pathCopy);
-                try {
-                  fileStore.fetch(pathCopy, getFrom);
-                } catch (Exception e) {
-                  log.error("Failed to download file: {}", pathCopy, e);
-                }
-                log.info("downloaded file: {}", pathCopy);
-              });
-      return response;
-    }
-
-    if (path == null) {
-      path = "";
-    }
-    FileStore.FileType typ = fileStore.getType(path, false);
-    if (typ == FileStore.FileType.NOFILE) {
-      final var fileMissingResponse = instantiateJerseyResponse(FileStoreMissingFileResponse.class);
-      fileMissingResponse.files = Collections.singletonMap(path, null);
-      return fileMissingResponse;
-    }
-    // TODO NOCOMMIT Will this serialize weirdly because Jersey doesn't know how to handle the
-    // 'FileDetails' objects in the list returned by FileStore.list()?  Or are they OK because
-    // they're MapWriter implementations?
-    if (typ == FileStore.FileType.DIRECTORY) {
-      final var directoryListingResponse =
-          instantiateJerseyResponse(FileStoreDirectoryListingResponse.class);
-      directoryListingResponse.files = Collections.singletonMap(path, fileStore.list(path, null));
-      return directoryListingResponse;
-    }
-    if (Boolean.TRUE.equals(meta)) {
-      if (typ == FileStore.FileType.FILE) {
-        int idx = path.lastIndexOf('/');
-        String fileName = path.substring(idx + 1);
-        String parentPath = path.substring(0, path.lastIndexOf('/'));
-        List<FileStore.FileDetails> l = fileStore.list(parentPath, s -> s.equals(fileName));
-
-        final var fileMetaResponse =
-            instantiateJerseyResponse(FileStoreDirectoryListingResponse.class);
-        fileMetaResponse.files = Collections.singletonMap(path, l.isEmpty() ? null : l.get(0));
-        return fileMetaResponse;
-      }
-    } else { // User wants to get the "raw" file
-      // TODO Should we be trying to json-ify otherwise "raw" files in this way?  It seems like a
-      // pretty sketchy idea, esp. for code with very little test coverage.  Consider removing
-      final var pathCopy2 = path;
-      if ("json".equals(req.getParams().get(CommonParams.WT))) {
-        final var jsonResponse = instantiateJerseyResponse(FileStoreJsonFileResponse.class);
-        try {
-          fileStore.get(
-              path,
-              it -> {
-                try {
-                  InputStream inputStream = it.getInputStream();
-                  if (inputStream != null) {
-                    jsonResponse.response = new String(inputStream.readAllBytes(), UTF_8);
-                  }
-                } catch (IOException e) {
-                  throw new SolrException(
-                      SolrException.ErrorCode.SERVER_ERROR, "Error reading file " + pathCopy2);
-                }
-              },
-              false);
-          return jsonResponse;
-        } catch (IOException e) {
-          throw new SolrException(
-              SolrException.ErrorCode.SERVER_ERROR, "Error getting file from path " + path);
-        }
-      } else {
-        ModifiableSolrParams solrParams = new ModifiableSolrParams();
-        solrParams.add(CommonParams.WT, FILE_STREAM);
-        req.setParams(SolrParams.wrapDefaults(solrParams, req.getParams()));
-        rsp.add(
-            FILE_STREAM,
-            (SolrCore.RawWriter)
-                os ->
-                    fileStore.get(
-                        pathCopy2,
-                        it -> {
-                          try {
-                            InputStream inputStream = it.getInputStream();
-                            if (inputStream != null) {
-                              inputStream.transferTo(os);
-                            }
-                          } catch (IOException e) {
-                            throw new SolrException(
-                                SolrException.ErrorCode.SERVER_ERROR,
-                                "Error reading file " + pathCopy2);
-                          }
-                        },
-                        false));
-      }
-    }
     return response;
   }
