+++ solr/core/src/java/org/apache/solr/handler/api/V2ApiUtils.java
@@ -19,6 +19,7 @@ package org.apache.solr.handler.api;
 
 import static org.apache.solr.client.solrj.impl.BinaryResponseParser.BINARY_CONTENT_TYPE_V2;
 import static org.apache.solr.common.params.CommonParams.WT;
+import static org.apache.solr.handler.ReplicationHandler.FILE_STREAM;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -30,6 +31,7 @@ import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.common.util.StrUtils;
 import org.apache.solr.common.util.Utils;
+import org.apache.solr.response.RawResponseWriter;
 import org.apache.solr.response.SolrQueryResponse;
 
 /** Utilities helpful for common V2 API declaration tasks. */
@@ -100,6 +102,8 @@ public class V2ApiUtils {
         return "application/xml";
       case "javabin":
         return BINARY_CONTENT_TYPE_V2;
+      case FILE_STREAM:
+        return RawResponseWriter.CONTENT_TYPE;
       default:
         return defaultMediaType;
     }
